Updated FMARCoreView to be responsible only by AR management
