+++ FantasmoSDK/app/src/main/res/layout/fmstatisticsview.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content">
 
@@ -103,8 +104,8 @@
         android:background="#BFFFFFFF"
         android:text="@string/translation"
         android:textColor="#000000"
-        app:layout_constraintStart_toStartOf="@+id/localizeTimeText"
-        app:layout_constraintTop_toBottomOf="@+id/localizeTimeText" />
+        app:layout_constraintStart_toStartOf="@+id/deviceLocationText"
+        app:layout_constraintTop_toBottomOf="@+id/deviceLocationText" />
 
     <TextView
         android:id="@+id/translationTextView"
@@ -434,4 +435,25 @@
         app:layout_constraintBottom_toBottomOf="@+id/featuresText"
         app:layout_constraintStart_toEndOf="@+id/featuresText"
         app:layout_constraintTop_toTopOf="@+id/featuresText" />
+
+    <TextView
+        android:id="@+id/deviceLocationText"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="4dp"
+        android:background="#BFFFFFFF"
+        android:text="@string/device_location"
+        android:textColor="@color/black"
+        app:layout_constraintStart_toStartOf="@+id/localizeTimeText"
+        app:layout_constraintTop_toBottomOf="@+id/localizeTimeText" />
+
+    <TextView
+        android:id="@+id/deviceLocationTextView"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:background="#BFFFFFFF"
+        android:textColor="@color/black"
+        app:layout_constraintBottom_toBottomOf="@+id/deviceLocationText"
+        app:layout_constraintStart_toEndOf="@+id/deviceLocationText"
+        app:layout_constraintTop_toTopOf="@+id/deviceLocationText" />
 </androidx.constraintlayout.widget.ConstraintLayout>
